Move getEntityInfo to its own service
