+++ packages/api/cms-api/src/common/entityInfo/entity-info.module.ts
@@ -1,0 +1,10 @@
+import { Module } from "@nestjs/common";
+
+import { EntityInfoService } from "./entity-info.service";
+
+@Module({
+    imports: [],
+    providers: [EntityInfoService],
+    exports: [EntityInfoService],
+})
+export class EntityInfoModule {}
