+++ packages/api/cms-api/src/dam/files/files-entity-info.service.ts
@@ -1,6 +1,6 @@
 import { forwardRef, Inject, Injectable } from "@nestjs/common";
 
-import { EntityInfoServiceInterface } from "../../dependencies/decorators/entity-info.decorator";
+import { EntityInfoServiceInterface } from "../../common/entityInfo/entity-info.decorator";
 import { FileInterface } from "./entities/file.entity";
 import { FilesService } from "./files.service";
