+++ packages/api/cms-api/src/dependencies/dependencies.service.ts
@@ -1,12 +1,10 @@
 import { InjectRepository } from "@mikro-orm/nestjs";
 import { AnyEntity, Connection, EntityManager, EntityRepository, Knex, QueryOrder } from "@mikro-orm/postgresql";
-import { Injectable, Logger, Type } from "@nestjs/common";
-import { INJECTABLE_WATERMARK } from "@nestjs/common/constants";
-import { ModuleRef } from "@nestjs/core";
+import { Injectable, Logger } from "@nestjs/common";
 import { subMinutes } from "date-fns";
 import { v4 as uuid } from "uuid";
 
-import { EntityInfoGetter, EntityInfoServiceInterface } from "./decorators/entity-info.decorator";
+import { EntityInfoService } from "../common/entityInfo/entity-info.service";
 import { DiscoverService } from "./discover.service";
 import { BaseDependencyInterface } from "./dto/base-dependency.interface";
 import { DependencyFilter, DependentFilter } from "./dto/dependencies.filter";
@@ -28,8 +26,8 @@ export class DependenciesService {
     constructor(
         @InjectRepository(BlockIndexRefresh) private readonly refreshRepository: EntityRepository<BlockIndexRefresh>,
         private readonly discoverService: DiscoverService,
+        private readonly entityInfoService: EntityInfoService,
         private entityManager: EntityManager,
-        private readonly moduleRef: ModuleRef,
     ) {
         this.connection = entityManager.getConnection();
     }
@@ -202,7 +200,7 @@ export class DependenciesService {
 
             let dependency: Dependency = result;
             if (instance) {
-                const entityInfo = await this.getEntityInfo(instance);
+                const entityInfo = await this.entityInfoService.getEntityInfo(instance);
                 dependency = { ...dependency, ...entityInfo };
             }
             ret.push(dependency);
@@ -244,7 +242,7 @@ export class DependenciesService {
 
             let dependency: Dependency = result;
             if (instance) {
-                const entityInfo = await this.getEntityInfo(instance);
+                const entityInfo = await this.entityInfoService.getEntityInfo(instance);
                 dependency = { ...dependency, ...entityInfo };
             }
             ret.push(dependency);
@@ -254,31 +252,6 @@ export class DependenciesService {
         const totalCount = countResult[0]?.count ?? 0;
 
         return new PaginatedDependencies(ret, Number(totalCount));
-    }
-
-    private async getEntityInfo(instance: object) {
-        const entityInfoGetter: EntityInfoGetter | undefined = Reflect.getMetadata(`data:entityInfo`, instance.constructor);
-
-        if (entityInfoGetter === undefined) {
-            this.logger.warn(
-                `Warning: ${instance.constructor.name} doesn't provide any entity info. You should add a @EntityInfo() decorator to the class. Otherwise it won't be displayed correctly as a dependency.`,
-            );
-            return {};
-        }
-
-        if (this.isService(entityInfoGetter)) {
-            const service = this.moduleRef.get(entityInfoGetter, { strict: false });
-            const { name, secondaryInformation } = await service.getEntityInfo(instance);
-            return { name, secondaryInformation };
-        } else {
-            const { name, secondaryInformation } = await entityInfoGetter(instance);
-            return { name, secondaryInformation };
-        }
-    }
-
-    private isService(entityInfoGetter: EntityInfoGetter): entityInfoGetter is Type<EntityInfoServiceInterface> {
-        // Check if class has @Injectable() decorator -> if true it's a service class else it's a function
-        return Reflect.hasMetadata(INJECTABLE_WATERMARK, entityInfoGetter);
     }
 
     private getQueryBuilderWithFilters(
